nothing new

+++ commands/clear.js
@@ -5,9 +5,11 @@ module.exports = {
     name: 'clear', 
     description: "clear commands",
     execute(message, args){
-if (!args[1]) return message.reply('How many messeges do u want me to sweep?')
-message.channel.bulkDelete(args[1]);
-
+        if (!args[1]) return message.reply('How many messeges do u want me to sweep?')
+        if (args[1] > 100) return message.reply ("I can't delete more than 100 mesaages dum dum")
+        message.channel.bulkDelete(args[1]);
+        message.channel.send ('Successully deleted '+(args[1]) + ' messages :D  got the order from => **' + message.author.username + '**')
+        
 
 
     }
